:recycle: dll split into .hpp and .cpp
